Update blog section and other fixes
