+++ Models/SocialMediaLink.cs
@@ -1,0 +1,7 @@
+namespace christian80gabi.github.io.Models;
+
+public class SocialMediaLink
+{
+    public string Title { get; set; } = string.Empty;
+    public string Link { get; set; } = string.Empty;
+}
