+++ Pages/MarkdownEditor.razor
@@ -14,6 +14,10 @@
     </div>
 </div>
 
+<footer class="footer mt-auto py-3 bg-body-tertiary">
+    <SocialMedia />
+</footer>
+
 
 @code {
     private string markdown { get; set; } = "";
