+++ Pages/Services.razor
@@ -7,23 +7,25 @@
 <div class="row mb-2">
     <div class="col-md-6">
         <div class="row g-0 border rounded mb-4 shadow-sm h-md-250 p-4">
-            <h2 class="border-3 text-primary border-secondary border-start ps-2">Web & Mobile Development</h2>
+            <h2 class="text-dark border-3 border-danger border-start ps-2">Web Development</h2>
+        </div>
+    </div>
+
+    <div class="col-md-6">
+        <div class="row g-0 border rounded mb-4 shadow-sm h-md-250 p-4">
+            <h2 class="text-secondary border-3 border-primary border-start ps-2">(ERP) Information System Analysis</h2>
         </div>
     </div>
     
     <div class="col-md-6">
         <div class="row g-0 border rounded mb-4 shadow-sm h-md-250 p-4">
-            <h2 class="border-3 border-warning border-start ps-2" style="color: #6f2eea;">Graphic Design</h2>
+            <h2 class="border-3 text-primary border-secondary border-start ps-2">WordPress</h2>
         </div>
     </div>
+    
     <div class="col-md-6">
         <div class="row g-0 border rounded mb-4 shadow-sm h-md-250 p-4">
-            <h2 class="text-dark border-3 border-danger border-start ps-2">CMS</h2>
-        </div>
-    </div>
-    <div class="col-md-6">
-        <div class="row g-0 border rounded mb-4 shadow-sm h-md-250 p-4">
-            <h2 class="text-success border-3 border-info border-start ps-2">Information System Analysis</h2>
+            <h2 class="border-3 border-warning border-start ps-2">Graphic Design</h2>
         </div>
     </div>
 </div>
@@ -32,7 +34,7 @@
     <h3><span class="text-secondary me-2">#</span> Summary</h3>
 </div>
 
-<div align="justify" class="font-monospace mt-3">
+<div class="font-monospace mt-3">
     <p>
         My technical expertise includes proficiency in various programming languages, including Java, Java EE, JavaScript, React, Angular, ASP.NET, Python, Django, and Spring Boot. Additionally, I possess a strong
         understanding of microservices and web services (SOAP, REST), as well as significant experience in data analysis
@@ -46,6 +48,19 @@
     </p>
 </div>
 
+<div class="col-12">
+        <div class="h-100 p-5 border rounded-3 text-bg-secondary text-white">
+            <h2>Languages of communication</h2>
+            <div>
+                <ul class="list-group list-group-flush">
+                    <li class="list-group-item text-bg-secondary text-white border-white">French (Native)</li>
+                    <li class="list-group-item text-bg-secondary text-white border-white">English (Native)</li>
+                    <li class="list-group-item text-bg-secondary text-white border-white">German (Conversationel)</li>
+                </ul>
+            </div>
+        </div>
+    </div>
+
 <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-3 pb-2 mb-3 border-bottom mt-4">
     <h3><span class="text-secondary me-2">#</span> Most Used Languages & Technologies</h3>
 </div>
@@ -54,7 +69,7 @@
     <div class="col-md-6">
         <div class="h-100 p-5 text-bg-dark rounded-3">
             <h2>Languages</h2>
-            <p>
+            <div>
                 <ul class="list-group list-group-flush">
                     <li class="list-group-item text-bg-dark border-white">Python</li>
                     <li class="list-group-item text-bg-dark border-white">JavaScript</li>
@@ -62,14 +77,14 @@
                     <li class="list-group-item text-bg-dark border-white">Java</li>
                     <li class="list-group-item text-bg-dark border-white">TypeScript</li>
                 </ul>
-            </p>
+            </div>
             
         </div>
     </div>
     <div class="col-md-6">
         <div class="h-100 p-5 border rounded-3">
             <h2>Frameworks & Libraries</h2>
-            <p>
+            <div>
                 <ul class="list-group list-group-flush">
                     <li class="list-group-item">Django</li>
                     <li class="list-group-item">Bootstrap</li>
@@ -77,19 +92,19 @@
                     <li class="list-group-item">ReactJS</li>
                     <li class="list-group-item">.NET</li>
                 </ul>
-            </p>
+            </div>
         </div>
     </div>
     <div class="col-12">
         <div class="h-100 p-5 border rounded-3 text-white" style="background-color: #0a2364;">
             <h2>Technologies & More</h2>
-            <p>
+            <div>
                 <ul class="list-group list-group-flush">
                     <li class="list-group-item text-white border-white" style="background-color: #0a2364;">Docker</li>
                     <li class="list-group-item text-white border-white" style="background-color: #0a2364;">Git</li>
                     <li class="list-group-item text-white border-white" style="background-color: #0a2364;">NodeJS</li>
                 </ul>
-            </p>
+            </div>
         </div>
     </div>
 </div>
