+++ Shared/SocialMedia.razor
@@ -1,6 +1,7 @@
-﻿@inject HttpClient Http
+﻿@using christian80gabi.github.io.Models
+@inject HttpClient Http
 
-<div class="d-flex justify-content-evenly">
+<div class="d-flex justify-content-evenly mt-5">
     <a href="@LinkedIn" target="_blank" data-bs-toggle="tooltip"
        data-bs-placement="top" data-bs-title="LinkedIn">
         <img src="icons/linkedin.svg" alt="Linkedin" height="25" />
@@ -15,22 +16,17 @@
     </a>
 </div>
 
-<div class="text-center text-muted mt-2">
-    <strong>@Title</strong> 
-</div>
-
-<div class="text-center text-bg-dark py-3 mt-2">
-    <a rel="license" href="http://creativecommons.org/licenses/by-nc-nd/4.0/"><img alt="Creative Commons License" style="border-width:0" src="https://i.creativecommons.org/l/by-nc-nd/4.0/88x31.png" /></a>
-    This work is licensed under a <a class="text-decoration-none text-info" rel="license" href="http://creativecommons.org/licenses/by-nc-nd/4.0/">Creative Commons Attribution-NonCommercial-NoDerivatives 4.0 International License</a>.
+<div class="text-center text-muted py-3">
+    <a rel="license" href="http://creativecommons.org/licenses/by-nc-nd/4.0/"><img alt="Creative Commons License" style="border-width:0" src="https://i.creativecommons.org/l/by-nc-nd/4.0/88x31.png" /></a> <strong>@Title</strong> 
 </div>
 
 @code {
     // Demonstrates how a parent component can supply parameters
     private static DateTime currentYear = DateTime.Now;
 
-    public string? Behance { get; set; }
-    public string? LinkedIn { get; set; }
-    public string? GitHub { get; set; }
+    private string? Behance { get; set; }
+    private string? LinkedIn { get; set; }
+    private string? GitHub { get; set; }
 
 
     [Parameter]
@@ -71,11 +67,4 @@
             }
         }
     }
-
-    public class SocialMediaLink
-    {
-        public string Title { get; set; } = string.Empty;
-        public string Link { get; set; } = string.Empty;
-    }
-
 }
