+++ wwwroot/data/posts/2024.01.04.01.md
@@ -13,13 +13,13 @@ Linux's story starts in 1991 with a young Finnish student named **Linus Torvalds
 
 The beauty of Linux extends far beyond the operating system itself. It opens the door to a vast universe of free and open-source software alternatives for popular proprietary applications. Let's peek into some examples:
 
-![LibreOffice_Logo](./resources/svg/LibreOffice_Logo.svg "LibreOffice logo")
+![LibreOffice_Logo](assets/svg/LibreOffice_Logo.svg "LibreOffice logo")
 * **Office Suite**: Ditch the hefty price tag of Microsoft Office and explore powerful alternatives like LibreOffice or FreeOffice. They offer familiar features and excellent compatibility with existing documents.
 
-![Inkscape_Logo](./resources/svg/Inkscape_Logo.svg "Inkscape logo")
+![Inkscape_Logo](assets/svg/Inkscape_Logo.svg "Inkscape logo")
 * **Graphics Creation**: Unleash your creativity with free software like GIMP (image editing), Krita (digital painting), and Inkscape (vector graphics). They offer professional-grade tools without the professional-grade price tag.
 
-![Visual_Studio_Code_Logo](./resources/svg/Visual_Studio_Code_Logo.svg "Visual Studio Code logo")
+![Visual_Studio_Code_Logo](assets/svg/Visual_Studio_Code_Logo.svg "Visual Studio Code logo")
 * **Web Development**: Code like a pro with free IDEs like Visual Studio Code and WebStorm, or build stunning websites with WordPress and countless free plugins. The possibilities are endless!
 
 This is just a taster of the vast FOSS universe waiting to be explored. In a future blog post, we'll delve deeper into specific alternatives for popular software, showcasing the power and variety available at your fingertips.
@@ -28,19 +28,19 @@ This is just a taster of the vast FOSS universe waiting to be explored. In a fut
 
 As we mentioned earlier, Linux comes in various flavors called distributions (distros), each catering to different needs and preferences. Let's introduce some of the major players:
 
-![Debian_Logo](./resources/svg/Debian_Logo.svg "Debian logo")
+![Debian_Logo](assets/svg/Debian_Logo.svg "Debian logo")
 * **Debian**: The granddaddy of distros, known for its stability and rock-solid foundation. The famous Ubuntu and Linux Mint are based on it. Think of it as the wise sage of the Linux world.
 
-![Slackware_Logo](./resources/svg/Slackware_logo.svg "Slackware logo")
+![Slackware_Logo](assets/svg/Slackware_logo.svg "Slackware logo")
 * **Slackware**: The minimalist veteran, perfect for experienced users who like things simple and in control. Think of it as your wise mentor, teaching you the basics of customization.
 
-![OpenSUSE_Logo](./resources/svg/OpenSUSE_Logo.svg "OpenSUSE logo")
+![OpenSUSE_Logo](assets/svg/OpenSUSE_Logo.svg "OpenSUSE logo")
 * **OpenSUSE**: A user-friendly distro with a focus on cutting-edge technologies and rolling releases (meaning you get the latest updates as soon as they're available). That's more about **Tumbleweed**, there is also **Leap** for more stability. Think of it as the adventurous explorer, always pushing the boundaries.
 
-![Fedora_Logo](./resources/svg/Fedora_logo.svg "Fedora logo")
+![Fedora_Logo](assets/svg/Fedora_logo.svg "Fedora logo")
 * **Fedora**: The community-driven innovator, known for its cutting-edge technologies and frequent releases. Think of it as the tech trendsetter, always ahead of the curve.
 
-![Arch_Linux_Logo](./resources/svg/Archlinux_Logo_Dark.svg "Arch Linux logo")
+![Arch_Linux_Logo](assets/svg/Archlinux_Logo_Dark.svg "Arch Linux logo")
 * **Arch Linux**: A DIY-focused distro for experienced users who enjoy customizing their systems to the finest detail. Think of it as your tinkerer's haven, where every piece is under your command.
 
 ---
@@ -48,7 +48,7 @@ As we mentioned earlier, Linux comes in various flavors called distributions (di
 > **Honorable mentions**: 
 > * **Gentoo**: A bit like Arch but for more experienced users,
 > * **Puppy Linux**: Lightweight, 
-> * **NixOS**: Easy reproductible system from one file, 
+> * **NixOS**: Easy reproducible system from one file, 
 > * **Void Linux**: Uh... the *void* long story short `(° o°)`.
 
 ----
@@ -63,7 +63,7 @@ However, it's important to remember that every user has different needs and pref
 
 ### OpenSUSE Tumbleweed: My Rolling Love Affair
 
-![OpenSUSE_Tumbleweed_Logo](./resources/svg/OpenSUSE_Tumbleweed_Green_Logo.svg "OpenSUSE Tumbleweed logo")
+![OpenSUSE_Tumbleweed_Logo](assets/svg/OpenSUSE_Tumbleweed_Green_Logo.svg "OpenSUSE Tumbleweed logo")
 
 As a Linux enthusiast myself, I'm smitten with OpenSUSE Tumbleweed. This rolling release distro keeps you on the cutting edge, delivering the latest software and bug fixes as they're ready. It's perfect for those who crave a dynamic computing experience and want to be at the forefront of technology.
 
@@ -75,6 +75,16 @@ Linux is no longer the obscure OS for geeks and hackers. It's a mature, powerful
 
 Remember, the beauty of Linux lies in its diversity and flexibility. Explore different distros, find one that resonates with you, and customize it to your heart's content. The possibilities are endless, and the adventure has just begun! Whether you're a seasoned.
 
-<style>
-    img { height: 100px; }
-</style>
+<div hidden>
+    <style>
+        img[alt=LibreOffice_Logo] { width: 100px; }
+        img[alt=Inkscape_Logo] { width: 100px; }
+        img[alt=Visual_Studio_Code_Logo] { width: 100px; }
+        img[alt=Debian_Logo] { width: 100px; }
+        img[alt=Slackware_Logo] { width: 100px; }
+        img[alt=OpenSUSE_Logo] { width: 100px; }
+        img[alt=Fedora_Logo] { width: 100px; }
+        img[alt=Arch_Linux_Logo] { width: 100px; }
+        img[alt=OpenSUSE_Tumbleweed_Logo] { width: 100px; }
+    </style>
+</div>
