+++ wwwroot/data/posts/2024.01.04.02.md
@@ -3,7 +3,7 @@
     By Christian G. N. (@christian80gabi)
     03/01/2023
 
-![Arch_Linux_Logo](./resources/svg/Archlinux_Logo_Dark.svg "Arch Linux logo")
+![Arch_Linux_Logo](assets/svg/Archlinux_Logo_Dark.svg "Arch Linux logo")
 
 Arch Linux. The name itself sends shivers down the spines of some, conjuring images of cryptic text files and hours spent wrestling with the command line. But what if I told you there was a way to experience the Arch magic without the steep learning curve? A streamlined installation process, guided by a friendly script, ready to go right on your Arch ISO? Introducing `archinstall`, the hidden gem waiting to be unleashed.
 
@@ -83,7 +83,3 @@ While Arch Linux still has its reputation for being a more hands-on distro, arch
 > **Bonus Tip**: For even more customization, explore the various archinstall profiles available online. These pre-configured profiles let you tailor your installation to specific needs, like gaming, development, or server setups.
 
 With archinstall, the gateway to Arch Linux is wider than ever. So, take the plunge and discover the beauty of Arch, simplified. [Arch Wiki](https://wiki.archlinux.org/)
-
-<style>
-    /* img[alt=Arch_Linux_Logo] { width: 100px; } */
-</style>
