+++ wwwroot/data/projects.json
@@ -5,23 +5,22 @@
     "status": "Done",
     "scope": "Personal",
     "source": "https://github.com/christian80gabi/PiWord",
-    "summary": "A Cross-platfom Word Discovery application using .NET Multi-platform App UI (.NET MAUI)."
+    "summary": "A Cross-platform Word Discovery application using .NET Multi-platform App UI (.NET MAUI)."
   },
   {
     "date": "2023-02-06",
-    "title": "Personal Portfollio",
-    "status": "Done",
+    "title": "Personal Portfolio",
+    "status": "In Progress",
     "scope": "Personal",
     "source": "https://christian80gabi.github.io/",
-    "summary": "Build my personal portfollio using Blazor and GitHub Pages."
+    "summary": "Build my personal portfolio using Blazor and GitHub Pages."
   },
   {
-    "date": "2023-02-01",
     "title": "Action",
     "status": "New",
     "scope": "Personal",
     "source": "",
-    "summary": "A Cross-platfom To-Do application using .NET Multi-platform App UI (.NET MAUI)."
+    "summary": "A Cross-platform To-Do application using .NET Multi-platform App UI (.NET MAUI)."
   },
   {
     "date": "2022-04-11",
@@ -29,7 +28,7 @@
     "status": "In Progress",
     "scope": "Open Source",
     "source": "https://github.com/christian80gabi/fullcalendar_django",
-    "summary": "An Agenda Django Web App using FullcalendarJS Library."
+    "summary": "A Django Web App using FullcalendarJS Library."
   },
   {
     "date": "2021-04-01",
@@ -37,6 +36,13 @@
     "status": "In Progress",
     "scope": "Contract",
     "source": "",
-    "summary": "A HRM (Human Resources Management System) by IDM Systems."
+    "summary": "A HRM (Human Resources Management System) by IDM Systems, Tunis, Tunisia."
+  },
+  {
+    "title": "Other",
+    "status": "In Progress",
+    "scope": "Multiple",
+    "source": "https://github.com/christian80gabi?tab=repositories",
+    "summary": "Check out."
   }
 ]
